requirements-pi.txt
numpy==1.26.4
pandas==2.2.2
ultralytics
flake8
flask
xgboost
pyserial
smbus2
psycopg2-binary
RPi.GPIO
adafruit-circuitpython-ads1x15
picamera2